alzheimers_vl_jepa/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
tqdm>=4.64.0
Pillow>=9.4.0
PyYAML>=6.0
